+++ CHANGELOG.md
@@ -9,4 +9,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Getting user input
+- Getting user input.
+- Enable raw mode in the terminal to make sure the input is valid straight-away.